Moved rate.Limiter into its own package
