+++ socketserver/server/rate/ratelimit_test.go
@@ -1,4 +1,4 @@
-package server
+package rate
 
 import (
 	"testing"
